Improve coverage of restricted interaction test views
This:

- adds a few more cases of non-restricted users accessing the
interactions API
- fixes a bug in a test that did not use the right authentication
when testing restricted interactions API
